Move default contents to separate files
