+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/composer/DefaultCenterContent.kt
@@ -1,0 +1,53 @@
+package io.getstream.chat.android.ui.message.composer
+
+import android.content.Context
+import android.text.Editable
+import android.util.AttributeSet
+import android.widget.FrameLayout
+import androidx.core.view.isVisible
+import com.google.android.material.internal.TextWatcherAdapter
+import io.getstream.chat.android.common.state.MessageInputState
+import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
+import io.getstream.chat.android.ui.databinding.StreamUiMessageComposerDefaultCenterContentBinding
+
+internal class DefaultCenterContent : FrameLayout {
+    var onTextChangedListener: (String) -> Unit = {}
+    var onClearButtonClickListener: () -> Unit = {}
+
+    private lateinit var binding: StreamUiMessageComposerDefaultCenterContentBinding
+
+    constructor(context: Context) : this(context, null)
+
+    constructor(context: Context, attrs: AttributeSet?) : this(context, attrs, 0)
+
+    constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int) : super(context,
+        attrs,
+        defStyleAttr) {
+        init()
+    }
+
+    private fun init() {
+        binding = StreamUiMessageComposerDefaultCenterContentBinding.inflate(streamThemeInflater, this)
+        binding.messageEditText.addTextChangedListener(object : TextWatcherAdapter() {
+            override fun afterTextChanged(s: Editable) {
+                onTextChangedListener(s.toString())
+            }
+        })
+        binding.clearCommandButton.setOnClickListener {
+            onClearButtonClickListener()
+        }
+    }
+
+    fun renderState(state: MessageInputState) {
+        val isClearInputButtonVisible = state.inputValue.isNotEmpty()
+        binding.clearCommandButton.isVisible = isClearInputButtonVisible
+
+        binding.messageEditText.apply {
+            val currentValue = text.toString()
+            val newValue = state.inputValue
+            if (newValue != currentValue) {
+                setText(state.inputValue)
+            }
+        }
+    }
+}
